Make luac.lua work on 64-Bit Machines
luac.lua didn't produce a correct bytecode on 64-Bit computers,
because it assumed that size_t is 4 bytes long.

The current solution assumes that uname -m returns a string that
contains 64 if it is run on a 64-Bit computer. This is currently
on true on x86_64, ppc64 and sparc64, but not on alpha. Moreover,
the kernel could run in 32-Bit mode and user space applications,
like lua, could run in 64-Bit mode. I found that is common on x86_64
machines with Mac OS X.

Testing the machine architecture with the following programme doesn't
work either, because the word size which lua was compiled with is
deciding: int main() { return sizeof(size_t) == 8 ? 1 : 0 }

The only solution would be to examine the lua binary, but this is
also not portable.

A patch for luac that adds preloading support exists [1], but has
not been merged into lua so far.

[1] http://lua-users.org/lists/lua-l/2009-10/msg00682.html