Test for 64-Bits with the compiler instead of uname
The previous solution used uname -m to determine whether luac.lua
should compile the bytecode for a 64-Bit machine. This not very
portable and problematic if the kernel runs in 32-Bit mode, but the
user-space applications run in 64-Bit mode.

The new solution assumes that gazelle is compiled for the same machine
architecture as lua was. This assumption is more likely.